Prevent players respawning at vanilla spawn in the overworld.
If a Player stands inside the block containing an EasySigns sleep sign and sets
his bed spawn location, the game will consider the bed to be obstructed (by the
sign) and clear the bed spawn to null.  The Player will then respawn at the
default, vanilla minecraft spawn location in the overworld.  That gives players
an exploitable warp to a central buff via the minigames area in the End.

This pull fixes the issue by detecting when the bed spawn is cleared and setting
it back to the default, non-team spawn in the End.  At worst, a player will
have to walk back to whatever minigame they were playing.