Highlight team member names in death messages, team chat and global chat.
Team members are highlighted in their team's respective chat color, regardless
of whether they are currently online.

Player names within words are not highlighted because they might refer to
a different player. For example, if a player named Fred is on the team,
then "hi fred" would be highlighted, but "hi alfred" would not.